Moves innards of organize imports into separate class.
